[Updated] Crafting process [Added] Novice Medic schematics
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@772 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/scene/SceneObjectImplementation.h
@@ -447,7 +447,7 @@ public:
 
 	inline float getPrecision(float num, int digits) {
 		float power = pow(10, digits);
-		return floor(num * power + .05f) / power;
+		return float(floor(num * power + .05f) / power);
 	}
 
 	inline uint64 getAssociatedArea() {
